select partially tested

+++ display.h
@@ -1,10 +1,10 @@
 #ifndef _DISPLAY_
 #define _DISPLAY_
 
-#define NBLANK 15
-#include <stdio.h>
-#include <stdlib.h>
+#define NBLANK (int)(15)
+#include <cstdio>
 #include <cstdlib>
+#include <cstring>
 #include "query_result.h"
 #include "types.h"
 #include "database.h"
